move main_section to project_components
