+++ apps/xim2_web/lib/xim2_web/components/layouts/app.html.heex
@@ -1,7 +1,5 @@
 <header class="px-4 sm:px-6 lg:px-8">Xim 2</header>
 <main class="px-4 sm:px-6 lg:px-8">
-  <div class="">
-    <.flash_group flash={@flash} />
-    <%= @inner_content %>
-  </div>
+  <.flash_group flash={@flash} />
+  <%= @inner_content %>
 </main>
