+++ apps/xim2_web/lib/xim2_web/live/monitor_live/index.ex
@@ -5,6 +5,7 @@ defmodule Xim2Web.MonitorLive.Index do
 
   alias Phoenix.PubSub
 
+  import Xim2Web.ProjectComponents
   import Xim2Web.Monitor.Components
 
   @items 500
